+++ network/p2p/readSubscription.go
@@ -2,11 +2,9 @@ package p2p
 
 import (
 	"context"
-	"fmt"
 	"strings"
 	"sync"
 
-	lcrypto "github.com/libp2p/go-libp2p-core/crypto"
 	"github.com/libp2p/go-libp2p-core/peer"
 	pubsub "github.com/libp2p/go-libp2p-pubsub"
 
@@ -103,43 +101,4 @@ func (r *readSubscription) receiveLoop(wg *sync.WaitGroup) {
 		// call the callback
 		r.callback(&msg, pid)
 	}
-}
-
-// messagePubKey extracts the public key of the envelope signer from a libp2p message.
-// The location of that key depends on the type of the key, see:
-// https://github.com/libp2p/specs/blob/master/peer-ids/peer-ids.md
-// This reproduces the exact logic of the private function doing the same decoding in libp2p:
-// https://github.com/libp2p/go-libp2p-pubsub/blob/ba28f8ecfc551d4d916beb748d3384951bce3ed0/sign.go#L77
-func messageSigningID(m *pubsub.Message) (peer.ID, error) {
-	var pubk lcrypto.PubKey
-
-	// m.From is the original sender of the message (versus `m.ReceivedFrom` which is the last hop which sent us this message)
-	pid, err := peer.IDFromBytes(m.From)
-	if err != nil {
-		return "", err
-	}
-
-	if m.Key == nil {
-		// no attached key, it must be extractable from the source ID
-		pubk, err = pid.ExtractPublicKey()
-		if err != nil {
-			return "", fmt.Errorf("cannot extract signing key: %s", err.Error())
-		}
-		if pubk == nil {
-			return "", fmt.Errorf("cannot extract signing key")
-		}
-	} else {
-		pubk, err = lcrypto.UnmarshalPublicKey(m.Key)
-		if err != nil {
-			return "", fmt.Errorf("cannot unmarshal signing key: %s", err.Error())
-		}
-
-		// verify that the source ID matches the attached key
-		if !pid.MatchesPublicKey(pubk) {
-			return "", fmt.Errorf("bad signing key; source ID %s doesn't match key", pid)
-		}
-	}
-
-	// the pid either contains or matches the signing pubKey
-	return pid, nil
 }
